packages/mcp/pyproject.toml
[project]
name = "reconly-mcp"
version = "1.0.0"
description = "Reconly MCP Server - Model Context Protocol integration for knowledge base access"
requires-python = ">=3.10"
authors = [
    { name = "Reconly Team", email = "hello@reconly.eu" }
]
license = { text = "AGPL-3.0" }
readme = "README.md"

dependencies = [
    "mcp>=1.0.0",
    "reconly-core>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
reconly-mcp = "reconly_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["reconly_mcp"]
